Correct a HybridStorage.Shares.cs (code line 55 and 68)

+++ src/CoiniumServ/Persistance/Layers/Hybrid/HybridStorage.Shares.cs
@@ -52,7 +52,7 @@ namespace CoiniumServ.Persistance.Layers.Hybrid
                 // add the share to round 
                 var currentKey = $"{_coin}:shares:round:current";
                 var miner = (IStratumMiner)share.Miner;
-                _redisProvider.Client.HIncrByFloat(currentKey, miner.Username, share.Difficulty);
+                _redisProvider.Client.HIncrByFloat(currentKey, miner.Username, (double)share.Difficulty);
                 //                _redisProvider.Client.HIncrByFloat(currentKey, share.Miner.Username, share.Difficulty);
 
                 // increment shares stats.
@@ -65,7 +65,7 @@ namespace CoiniumServ.Persistance.Layers.Hybrid
                     var hashrateKey = $"{_coin}:hashrate";
                     var randomModifier = Convert.ToString(miner.ValidShareCount, 16).PadLeft(8, '0');
                     string modifiedUsername = miner.Username + randomModifier;
-                    var entry = string.Format("{0}:{1}", share.Difficulty, modifiedUsername);
+                    var entry = string.Format("{0}:{1}", (double)share.Difficulty, modifiedUsername);
                     _redisProvider.Client.ZAdd(hashrateKey, Tuple.Create((double)TimeHelpers.NowInUnixTimestamp(), entry));
                     //                    var entry = $"{share.Difficulty}:{share.Miner.Username}";
                     //_redisProvider.Client.ZAdd(hashrateKey, Tuple.Create((double)TimeHelpers.NowInUnixTimestamp(), entry));
